Move benches to separate folder
