+++ benches/readhex.rs
@@ -1,0 +1,14 @@
+#![feature(test)]
+
+extern crate indolenjson;
+extern crate test;
+
+use indolenjson::readhex::*;
+use test::{black_box, Bencher};
+
+#[bench]
+fn read_hexdigit(b: &mut Bencher) {
+    let x = black_box(b"0123");
+    b.bytes = x.len() as u64;
+    b.iter(|| { read_hexdigits(x[0], x[1], x[2], x[3]) });
+}
