+++ src/lib.rs
@@ -1,6 +1,2 @@
-#![feature(test)]
-
-extern crate test;
-
 pub mod compact;
 pub mod readhex;
